chore(release): support release crds and deploy files
