+++ charts/tidb-operator/Chart.yaml
@@ -1,8 +1,6 @@
-apiVersion: v1
+apiVersion: v2
 description: tidb-operator Helm chart for Kubernetes
 name: tidb-operator
-version: v2-canary
-appVersion: v2-canary
 home: https://github.com/pingcap/tidb-operator
 sources:
   - https://github.com/pingcap/tidb-operator
@@ -20,3 +18,9 @@ maintainers:
     email: liubo02@pingcap.com
   - name: fgksgf
     email: jianghuaxi@pingcap.com
+
+kubeVersion: ">=1.30.0"
+
+# will be set when package
+version: v0.0.0
+appVersion: v0.0.0
