+++ charts/tidb-operator/templates/deployment.yaml
@@ -9,10 +9,8 @@ metadata:
   namespace: {{ .Release.Namespace }}
   labels:
     app.kubernetes.io/name: {{ template "chart.name" . }}
-    app.kubernetes.io/managed-by: {{ .Release.Service }}
     app.kubernetes.io/instance: {{ .Release.Name }}
     app.kubernetes.io/component: tidb-operator
-    helm.sh/chart: {{ .Chart.Name }}-{{ .Chart.Version | replace "+"  "_" }}
 spec:
   replicas: {{ .Values.operator.replicas }}
   selector:
@@ -44,7 +42,7 @@ spec:
     {{- end }}
       containers:
       - command:
-        - /operator
+        - /tidb-operator
         args:
           - --backup-manager-image={{ .Values.backupManager.image }}
         {{- if .Values.operator.extraArgs }}
