+++ charts/tidb-operator/templates/rbac.yaml
@@ -10,10 +10,8 @@ metadata:
   namespace: {{ .Release.Namespace }}
   labels:
     app.kubernetes.io/name: {{ template "chart.name" . }}
-    app.kubernetes.io/managed-by: {{ .Release.Service }}
     app.kubernetes.io/instance: {{ .Release.Name }}
     app.kubernetes.io/component: tidb-operator
-    helm.sh/chart: {{ .Chart.Name }}-{{ .Chart.Version | replace "+"  "_" }}
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
@@ -21,31 +19,32 @@ metadata:
   name: {{ .Release.Name }}:tidb-operator
   labels:
     app.kubernetes.io/name: {{ template "chart.name" . }}
-    app.kubernetes.io/managed-by: {{ .Release.Service }}
     app.kubernetes.io/instance: {{ .Release.Name }}
     app.kubernetes.io/component: tidb-operator
-    helm.sh/chart: {{ .Chart.Name }}-{{ .Chart.Version | replace "+"  "_" }}
 rules:
   - nonResourceURLs: ["/metrics"]
     verbs: ["get"]
   - apiGroups: [""]
-    resources: ["configmaps", "events", "persistentvolumeclaims", "pods", "services"]
-    verbs: ["*"]
+    resources: ["configmaps", "persistentvolumeclaims", "pods", "services"]
+    verbs: ["get", "list", "watch", "create", "delete", "update", "patch"]
+  - apiGroups: [""]
+    resources: ["events"]
+    verbs: ["create"]
   - apiGroups: [""]
     resources: ["nodes", "persistentvolumes", "secrets"]
     verbs: ["get", "list", "watch"]
   - apiGroups: ["apps"]
     resources: ["controllerrevisions"]
-    verbs: ["*"]
+    verbs: ["get", "list", "watch", "create", "delete", "update", "patch"]
   - apiGroups: ["batch"]
     resources: ["jobs"]
-    verbs: ["*"]
+    verbs: ["get", "list", "watch", "create", "delete", "update", "patch"]
   - apiGroups: ["storage.k8s.io"]
     resources: ["storageclasses", "volumeattributesclasses"]
     verbs: ["get", "list", "watch"]
   - apiGroups: ["coordination.k8s.io"]
     resources: ["leases"]
-    verbs: ["*"]
+    verbs: ["get", "list", "watch", "create", "delete", "update", "patch"]
   - apiGroups: ["core.pingcap.com"]
     resources: ["*"]
     verbs: ["*"]
@@ -59,10 +58,8 @@ metadata:
   name: {{ .Release.Name }}:tidb-operator
   labels:
     app.kubernetes.io/name: {{ template "chart.name" . }}
-    app.kubernetes.io/managed-by: {{ .Release.Service }}
     app.kubernetes.io/instance: {{ .Release.Name }}
     app.kubernetes.io/component: tidb-operator
-    helm.sh/chart: {{ .Chart.Name }}-{{ .Chart.Version | replace "+"  "_" }}
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: ClusterRole
