+++ charts/tidb-operator/templates/webhook.yaml
@@ -10,10 +10,8 @@ metadata:
   name: "delete-pod.policy.pingcap.com"
   labels:
     app.kubernetes.io/name: {{ template "chart.name" . }}
-    app.kubernetes.io/managed-by: {{ .Release.Service }}
     app.kubernetes.io/instance: {{ .Release.Name }}
     app.kubernetes.io/component: webhook
-    helm.sh/chart: {{ .Chart.Name }}-{{ .Chart.Version | replace "+"  "_" }}
 spec:
   failurePolicy: Fail
   matchConstraints:
@@ -32,10 +30,8 @@ metadata:
   name: "delete-pod.policy.pingcap.com"
   labels:
     app.kubernetes.io/name: {{ template "chart.name" . }}
-    app.kubernetes.io/managed-by: {{ .Release.Service }}
     app.kubernetes.io/instance: {{ .Release.Name }}
     app.kubernetes.io/component: webhook
-    helm.sh/chart: {{ .Chart.Name }}-{{ .Chart.Version | replace "+"  "_" }}
 spec:
   policyName: "delete-pod.policy.pingcap.com"
   validationActions: [Deny]
