--- hack/yaml.sh
@@ -1,38 +1,0 @@
-#!/usr/bin/env bash
-# Copyright 2024 PingCAP, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-set -o errexit
-set -o nounset
-set -o pipefail
-
-ROOT=$(cd $(dirname "${BASH_SOURCE[0]}")/..; pwd -P)
-source $ROOT/hack/lib/vars.sh
-
-OPERATOR=$ROOT/manifests/tidb-operator.yaml
-CRDS=$ROOT/manifests/tidb-operator.crds.yaml
-
-echo -n > $CRDS
-echo -n > $OPERATOR
-for f in $ROOT/manifests/crd/*.yaml; do
-    echo "append $f"
-    cat $f >> $CRDS
-    cat $f >> $OPERATOR
-done
-
-for f in $ROOT/manifests/deploy/*.yaml; do
-    echo "append $f"
-    cat $f >> $OPERATOR
-done
