--- manifests/deploy/02-role.yaml
@@ -1,178 +1,0 @@
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  name: tidb-operator
-rules:
-- nonResourceURLs:
-  - /metrics
-  verbs:
-  - get
-- apiGroups:
-  - ""
-  resources:
-  - configmaps
-  - events
-  - persistentvolumeclaims
-  - pods
-  - services
-  verbs:
-  - create
-  - delete
-  - get
-  - list
-  - patch
-  - update
-  - watch
-- apiGroups:
-  - ""
-  resources:
-  - nodes
-  - persistentvolumes
-  - secrets
-  verbs:
-  - get
-  - list
-  - watch
-- apiGroups:
-  - apps
-  resources:
-  - controllerrevisions
-  verbs:
-  - create
-  - delete
-  - get
-  - list
-  - patch
-  - update
-  - watch
-- apiGroups:
-  - batch
-  resources:
-  - jobs
-  verbs:
-  - create
-  - delete
-  - get
-  - list
-  - patch
-  - update
-  - watch
-- apiGroups:
-  - br.pingcap.com
-  resources:
-  - backups
-  - backupschedules
-  - compactbackups
-  - restores
-  verbs:
-  - get
-  - list
-  - update
-  - watch
-- apiGroups:
-  - br.pingcap.com
-  resources:
-  - backups/status
-  - backupschedules/status
-  - compactbackups/status
-  - restores/status
-  verbs:
-  - get
-  - patch
-  - update
-- apiGroups:
-  - coordination.k8s.io
-  resources:
-  - leases
-  verbs:
-  - create
-  - delete
-  - get
-  - list
-  - patch
-  - update
-  - watch
-- apiGroups:
-  - core.pingcap.com
-  resources:
-  - clusters
-  - pdgroups/status
-  - ticdcgroups/status
-  - tidbgroups/status
-  - tiflashgroups/status
-  - tikvgroups/status
-  - tsogroups/status
-  - schedulergroups/status
-  verbs:
-  - get
-  - list
-  - update
-  - watch
-- apiGroups:
-  - core.pingcap.com
-  resources:
-  - clusters/status
-  verbs:
-  - get
-  - patch
-  - update
-- apiGroups:
-  - core.pingcap.com
-  resources:
-  - pdgroups
-  - ticdcgroups
-  - tidbgroups
-  - tiflashgroups
-  - tikvgroups
-  - tsogroups
-  - schedulergroups
-  verbs:
-  - delete
-  - get
-  - list
-  - update
-  - watch
-- apiGroups:
-  - core.pingcap.com
-  resources:
-  - pds
-  - ticdcs
-  - tidbs
-  - tiflashes
-  - tikvs
-  - tsos
-  - schedulers
-  verbs:
-  - create
-  - delete
-  - get
-  - list
-  - patch
-  - update
-  - watch
-- apiGroups:
-  - core.pingcap.com
-  resources:
-  - pds/status
-  - ticdcs/status
-  - tidbs/status
-  - tiflashes/status
-  - tikvs/status
-  - tsos/status
-  - schedulers/status
-  verbs:
-  - get
-  - list
-  - patch
-  - update
-  - watch
-- apiGroups:
-  - storage.k8s.io
-  resources:
-  - storageclasses
-  - volumeattributesclasses
-  verbs:
-  - get
-  - list
-  - watch
