--- manifests/deploy/03-role_binding.yaml
@@ -1,13 +1,0 @@
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: tidb-operator
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: tidb-operator
-subjects:
-- kind: ServiceAccount
-  name: tidb-operator
-  namespace: tidb-admin
