--- manifests/deploy/05-deploy.yaml
@@ -1,52 +1,0 @@
----
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: tidb-operator
-  namespace: tidb-admin
-spec:
-  selector:
-    matchLabels:
-      app.kubernetes.io/name: tidb-operator
-  replicas: 1
-  template:
-    metadata:
-      labels:
-        app.kubernetes.io/name: tidb-operator
-    spec:
-      securityContext:
-        runAsNonRoot: true
-      containers:
-      - command:
-        - /operator
-        args:
-          - --backup-manager-image=pingcap/backup-manager:latest
-        image: pingcap/operator:latest
-        imagePullPolicy: IfNotPresent
-        name: operator
-        securityContext:
-          allowPrivilegeEscalation: false
-          capabilities:
-            drop:
-            - "ALL"
-        livenessProbe:
-          httpGet:
-            path: /healthz
-            port: 8081
-          initialDelaySeconds: 15
-          periodSeconds: 20
-        readinessProbe:
-          httpGet:
-            path: /readyz
-            port: 8081
-          initialDelaySeconds: 5
-          periodSeconds: 10
-        resources:
-          limits:
-            cpu: 500m
-            memory: 128Mi
-          requests:
-            cpu: 10m
-            memory: 64Mi
-      serviceAccountName: tidb-operator
-      terminationGracePeriodSeconds: 10
